refactor(src-tauri/src/commands/groups): expand groups commands to module with their own files
