+++ src-tauri/src/commands/groups/mod.rs
@@ -1,0 +1,19 @@
+mod create_group;
+mod delete_message;
+mod get_group;
+mod get_group_admins;
+mod get_group_and_messages;
+mod get_group_members;
+mod get_groups;
+mod rotate_key_in_group;
+mod send_mls_message;
+
+pub use create_group::create_group;
+pub use delete_message::delete_message;
+pub use get_group::get_group;
+pub use get_group_admins::get_group_admins;
+pub use get_group_and_messages::get_group_and_messages;
+pub use get_group_members::get_group_members;
+pub use get_groups::get_groups;
+pub use rotate_key_in_group::rotate_key_in_group;
+pub use send_mls_message::send_mls_message;
